// server/common-interfaces/src/model/st-market/st-exports.model.ts
import { STMarketChartData } from './st-market.model';
export interface STMarketExports {
    asia: STMarketChartData;
    europe: STMarketChartData;
    europeanUnion: STMarketChartData;
    france: STMarketChartData;
    unitedkingdom: STMarketChartData;
    latinAmerica: STMarketChartData;
    germany: STMarketChartData;
    japan: STMarketChartData;
    china: STMarketChartData;
    canada: STMarketChartData;
    timestamp: number[];
}